move read_file tests to own file
